% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_by_region.R
\name{average_by_region}
\alias{average_by_region}
\title{Find Average Starting Salary by Region}
\usage{
average_by_region(x)
}
\arguments{
\item{x, }{the data on salaries based on college}
}
\value{
information about the average salaries based on medians according to region
}
\description{
Find Average Starting Salary by Region
}
\examples{
    average_by_region()
}
