% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_by_type.R
\name{average_by_type}
\alias{average_by_type}
\title{Find Average Starting Salary by College Type}
\usage{
average_by_type(x)
}
\arguments{
\item{x, }{the data on salaries based on college}
}
\value{
information about the average salaries based on medians according to the kind of undergraduate school attended
}
\description{
Find Average Starting Salary by College Type
}
\examples{
    average_by_type()
}
