% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_college.R
\name{find_college}
\alias{find_college}
\title{Find College}
\usage{
find_college(x)
}
\arguments{
\item{x, }{the data of the universities}
}
\value{
information about the median salary of graduates
}
\description{
Find College
}
\examples{
    find_college("University of Mississippi")
}
