% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_degree.R
\name{find_degree}
\alias{find_degree}
\title{Find Degree}
\usage{
find_degree(x)
}
\arguments{
\item{x, }{the data of the degrees}
}
\value{
information about the median salary of graduates according to undergraduate degree
}
\description{
Find Degree
}
\examples{
    find_degree("Biology")
}
